backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
boto3
scikit-learn==1.6.1
pydantic
pandas
xgboost
python-dotenv
psycopg2-binary
sqlalchemy